move tests into StorageApi dir
